+++ plugins/classify/types/index.d.ts
@@ -2,11 +2,10 @@ import type { Plugin } from 'wtf_wikipedia'
 
 /** the result of doc.classify() */
 export type ClassifyResult = {
-  root?: string
+  root: string | null
   type: string | null
   score: number
-  details?: Record<string, unknown[]>
-  detail?: Record<string, unknown[]>
+  details: Record<string, unknown[]>
 }
 
 declare module 'wtf_wikipedia' {